Remove pipenv
Pipenv works extremely slow, and it doesn't tell the progress to me.
When I had installed pybrain and wanted to uninstall it, pipenv couldn't
lock dependencies after about an half hour so I aborted it.
So I work around by removing the virtualenv via `pipenv --rm` and then
installing the fresh virtualenv. The same problem happened when I tried
to install pytorch; I waited 30 minutes and aborted it.

So I removed pipenv and came back to the classic virtualenv (yeah, with
virtualenvwrapper). One of the advantages for me to use `pip` directly
is that it shows the installing progress for each package installation.
Pipenv does not inform me any specific when installing packages and
locking dependencies.

Also the direct `pip` is much faster. It has nothing to do with
dependency resolution; it actually is the point of pipenv, but the
problem is the speed. I agree that `requirements.txt` sucks though, I
think that it is just enough to record packages what I installed, for
this small scientific notebooks.