Add Section 11.6 with the partial implementation
Try only the sigmoid model with various numbers of hidden units.
btw, I couldn't figure out what the Bayes rate actually is, and why the
simulated test error is better than the Bayes rate (if the Bayes rate
is the variance of the error variable as the textbook said).
More to come!